preserve dimension as part of CoordinateArraySequence and while reading WKT (#345)
git-svn-id: http://svn.osgeo.org/geos/trunk@2990 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ tests/unit/noding/BasicSegmentStringTest.cpp
@@ -64,7 +64,7 @@ namespace tut
     template<>
     void object::test<1>()
     {
-    	CoordinateSequenceAutoPtr cs(csFactory->create(0, 2));
+    	CoordinateSequenceAutoPtr cs(csFactory->create((size_t)0, 2));
 
 	ensure(0 != cs.get());
 
@@ -107,7 +107,7 @@ namespace tut
     template<>
     void object::test<2>()
     {
-    	CoordinateSequenceAutoPtr cs(csFactory->create(0, 2));
+    	CoordinateSequenceAutoPtr cs(csFactory->create((size_t)0, 2));
 
 	ensure(0 != cs.get());
 
@@ -143,7 +143,7 @@ namespace tut
     template<>
     void object::test<3>()
     {
-    	CoordinateSequenceAutoPtr cs(csFactory->create(0, 2));
+    	CoordinateSequenceAutoPtr cs(csFactory->create((size_t)0, 2));
 
 	ensure(0 != cs.get());
